campus-guide-backend/requirements.txt
fastapi[all]
google-generativeai
python-dotenv